+ Checkpoint   + the MBE is functioning, to the second order, but energy is not correctly returned to the final report
 On branch manybody_expansion
 Your branch is up-to-date with 'origin/manybody_expansion'.

 Changes to be committed:
	modified:   ../../electron_correlation/client.cpp
	modified:   ../../electron_correlation/makefile
	modified:   ../../electron_correlation/setting.cpp
	modified:   ../../electron_correlation/setting.hpp
	modified:   ../../electron_correlation/test/main.cpp
	modified:   ../../file_system/directory.hpp
	modified:   ../../file_system/test/main.cpp
	modified:   ../../interface_to_third_party/external_program_config_base.hpp
	modified:   ../../interface_to_third_party/external_program_request.hpp
	modified:   ../../interface_to_third_party/makefile
	modified:   ../../interface_to_third_party/molpro_agent.cpp
	modified:   ../../interface_to_third_party/molpro_config.hpp
	modified:   ../agent.cpp
	modified:   ../compute_expansion_term_periodic_traits.hpp
	modified:   ../compute_interaction_energy_traits.hpp
	modified:   ../config.hpp
	modified:   ../makefile
	modified:   ../request.hpp
	modified:   main.cpp
	modified:   makefile
	modified:   ../../structure/lattice_parameter.hpp
	modified:   ../../structure/lattice_template.hpp

 Untracked files:
	../../../bin/
	../../chemistry/test/main.o
	../../chemistry/test/test_element
	../../electron_correlation/client.o
	../../electron_correlation/libiquads_electron_correlation.a
	../../electron_correlation/report.o
	../../electron_correlation/setting.o
	../../electron_correlation/test/client.o
	../../electron_correlation/test/input/
	../../electron_correlation/test/main.o
	../../electron_correlation/test/output/
	../../electron_correlation/test/report.o
	../../electron_correlation/test/scratch/
	../../electron_correlation/test/setting.o
	../../electron_correlation/test/test_electron_correlation
	../../file_system/test/main.o
	../../file_system/test/test_file_system
	../../interface_to_third_party/libiquads_third_party.a
	../../interface_to_third_party/molpro_agent.o
	../../interface_to_third_party/test/molpro_agent.o
	../../interface_to_third_party/test/test_molpro_agent
	../agent.o
	../client.o
	../config.o
	../iquads_manybody_expansion
	../libiquads_manybody_expansion.a
	../main.o
	../report.o
	../request.o
	.exrc
	main.o
	try/
	../../mbe_checkout.sh
	../../structure/test/main.o
	../../structure/test/test
	../../utility/test/main.o
	../../utility/test/test_utility
	../../../test/nonadiabatic_coupling/
	../../../test/numerical_nonadiabatic_coupling/

+++ src/interface_to_third_party/external_program_request.hpp
@@ -83,8 +83,8 @@ public:
   ExternalProgramRequest( molecule_name_type     molecule_name,
                           molecule_obj_type      molecule_obj,
                           basis_set_name_type    basis_set_name,
-                          size_t                 spin_value,
-                          size_t                 sym_value,
+                          int                    spin_value,
+                          int                    sym_value,
                           external_program_type  external_program,
                           mode_type              mode,
                           calculation_type       calculation,
@@ -107,9 +107,9 @@ public:
     { return this->basis_set_name_; }
   external_program_type external_program() const
     { return this->external_program_; }
-  size_t spin() const
+  int spin() const
     { return this->spin_; }
-  size_t sym() const 
+  int sym() const 
     { return this->sym_; }
   mode_type mode() const
     { return this->mode_; }
@@ -139,8 +139,8 @@ public:
 
 private:
   molecule_info_type      molecule_info_;
-  size_t                  spin_;
-  size_t                  sym_;
+  int                     spin_;
+  int                     sym_;
   basis_set_name_type     basis_set_name_;
   external_program_type   external_program_;
   mode_type               mode_;
